Introduce Criteria mapper
