+++ client-service/src/test/java/com/jaspersoft/android/sdk/service/repository/InternalSearchCriteriaTest.java
@@ -27,14 +27,7 @@ package com.jaspersoft.android.sdk.service.repository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-
 import junitparams.JUnitParamsRunner;
-import junitparams.Parameters;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -45,141 +38,6 @@ import static org.hamcrest.core.Is.is;
  */
 @RunWith(JUnitParamsRunner.class)
 public class InternalSearchCriteriaTest {
-
-    @Test
-    public void shouldIncludeCountInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .limit(101)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("limit", "101");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeOffsetInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .offset(100)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("offset", "100");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeRecursiveInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .recursive(true)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("recursive", "true");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeForceFullPageInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .forceFullPage(true)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("forceFullPage", "true");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeForceTotalCountPageInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .forceTotalCount(true)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("forceTotalCount", "true");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeQueryPageInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .query("any")
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("q", "any");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIncludeFolderUriInParams() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .folderUri("/")
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        resultMap.put("folderUri", "/");
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldIgnoreEmptyQuery() {
-        InternalCriteria criteria = InternalCriteria.builder()
-                .query("")
-                .create();
-
-        Map<String, Object> emptytMap = new HashMap<>();
-        assertThat(criteria.toMap(), is(emptytMap));
-    }
-
-    @Test
-    @Parameters({
-            "REPORT|reportUnit",
-            "DASHBOARD|dashboard",
-            "LEGACY_DASHBOARD|legacyDashboard",
-            "ALL|reportUnit:dashboard:legacyDashboard",
-            "REPORT:DASHBOARD|reportUnit:dashboard",
-    })
-    public void criteriaShouldIncludeTypeInParams(String flags, String types) throws Exception {
-        int mask = 0;
-        if (flags.contains(":")) {
-            for (String flag : flags.split(":")) {
-                mask |= (Integer) SearchCriteria.class.getField(flag).get(null);
-            }
-        } else {
-            mask = (Integer) SearchCriteria.class.getField(flags).get(null);
-        }
-
-        InternalCriteria criteria = InternalCriteria.builder()
-                .resourceMask(mask)
-                .create();
-
-        Map<String, Object> resultMap = new HashMap<>();
-        Set<String> typeSet = new HashSet<>();
-        if (types.contains(":")) {
-            typeSet.addAll(Arrays.asList(types.split(":")));
-        } else {
-            typeSet.add(types);
-        }
-        resultMap.put("type", typeSet);
-
-        assertThat(criteria.toMap(), is(resultMap));
-    }
-
-    @Test
-    public void shouldReturnEmptyParamsIfNoSupplied() {
-        InternalCriteria criteria = InternalCriteria.builder().create();
-        Map<String, Object> resultMap = new HashMap<>();
-        assertThat(criteria.toMap(), is(resultMap));
-    }
 
     @Test
     public void newBuilderShouldCopyCount() {
